HIVE-22347 Break up DDLSemanticAnalyzer - extract Other analyzers
