+++ ql/src/java/org/apache/hadoop/hive/ql/parse/AnalyzeCommandUtils.java
@@ -25,7 +25,14 @@ import org.apache.hadoop.hive.ql.exec.Utilities;
 import org.apache.hadoop.hive.ql.metadata.Table;
 import org.apache.hadoop.hive.ql.session.SessionState;
 
-public class AnalyzeCommandUtils {
+/**
+ * Utilities for semantic analyzers.
+ */
+public final class AnalyzeCommandUtils {
+  private AnalyzeCommandUtils() {
+    throw new UnsupportedOperationException("AnalyzeCommandUtils should not be instantiated");
+  }
+
   public static boolean isPartitionLevelStats(ASTNode tree) {
     boolean isPartitioned = false;
     ASTNode child = (ASTNode) tree.getChild(0);
@@ -50,7 +57,7 @@ public class AnalyzeCommandUtils {
     ASTNode child = ((ASTNode) tree.getChild(0).getChild(1));
     Map<String,String> partSpec = new HashMap<String, String>();
     if (child != null) {
-      partSpec = DDLSemanticAnalyzer.getValidatedPartSpec(tbl, child, hiveConf, false);
+      partSpec = BaseSemanticAnalyzer.getValidatedPartSpec(tbl, child, hiveConf, false);
     } //otherwise, it is the case of analyze table T compute statistics for columns;
     return partSpec;
   }
